Convert WAL id int to long.

+++ src/ZoneTree/Exceptions/WriteAheadLogCorruptionException.cs
@@ -4,7 +4,7 @@ public class WriteAheadLogCorruptionException : ZoneTreeException
 {
     public Dictionary<int, Exception> Exceptions { get; }
 
-    public WriteAheadLogCorruptionException(int segmentId, 
+    public WriteAheadLogCorruptionException(long segmentId, 
         Dictionary<int, Exception> exceptions)
         : base($"Write ahead log with segment id = {segmentId} is corrupted.")
     {
